DRY #1: move the hec to a container
